Remove Task and PatchOp; allow a patch without data
Task leaves dascore; another package builds tasks and workflows on top of
PatchProcessor. Delete Task, intern, @task, make_function_task_class,
save/load, PatchOp, fn.op(), the builtin Concatenate/Stack/Ufunc/ArrayFunc
tasks, the workflow tutorial, and the dascore.workflow package.

Survivors move to their final homes: core/processor.py (PatchProcessor,
PatchMeta, register_kernel, register_implementation), utils/identity.py,
utils/serialize.py, and utils/patch_registry.py (tag registry and
fingerprint_call, split out so utils/patch.py can hold the checks again
without an import cycle).

Every fingerprint now has one spelling, {"operation", "version",
"params"}, and a patch argument encodes under its own $patch tag rather
than by the module path of a private class. processing_id values change.

Patch(coords=..., dtype=...) builds a patch without data: dtype is
stored, .data and every patch function raise PatchDataError, new(data=)
fills it, and drop_data() makes one.
