+++ docs/contributing/extending_dascore.qmd
@@ -77,13 +77,13 @@ Plain NumPy-based functions may raise or return a NumPy-backed patch for other a
 
 NumPy functions such as `np.mean(patch)` and reductions over excluded dtypes always use this fallback; `__array__`-only objects report `numpy` as their backend.
 
-Some operations use a [`PatchProcessor`](`dascore.workflow.processor.PatchProcessor`) with separate metadata and array kernels. A backend package can register its own kernel without changing the operation:
+Some operations use a [`PatchProcessor`](`dascore.core.processor.PatchProcessor`) with separate metadata and array kernels. A backend package can register its own kernel without changing the operation:
 
 ```{python}
 import numpy as np
 
 from dascore.proc.basic import Abs
-from dascore.workflow import register_kernel
+from dascore.core.processor import register_kernel
 
 
 @register_kernel(Abs, "mybackend")
