+++ tests/test_proc/test_adaptive_spectral_filter.py
@@ -528,12 +528,11 @@ class TestAdaptiveSpectralFilter:
         assert out.data.dtype == np.float32
         assert np.isfinite(out.data).all()
 
-    def test_op_is_the_processor(self) -> None:
-        """The seam: the call names the processor, and the two routes agree."""
+    def test_the_processor_is_the_call(self) -> None:
+        """The seam: the processor and the method agree."""
         patch = _patch((64, 64), ("distance", "time"), dtype=np.float32)
-        op = adaptive_spectral_filter_func.op(time=16, distance=16, samples=True)
+        op = AdaptiveSpectralFilter(time=16, distance=16, samples=True)
 
-        assert isinstance(op, AdaptiveSpectralFilter)
         expected = patch.adaptive_spectral_filter(time=16, distance=16, samples=True)
         assert op(patch).equals(expected)
